Rewritten action scheduling algorithm
-- Summary

Actions are scheduled using an algorithm which operates on a proper
dependency graph, assigning choices so that there are no cycles.

We also change how component hashes are computed.  Now hashes are a
property of a component as a whole and not of a single build.
The recursive hash of a component depends on the set of all the
components that can be reached recursively from the dependency graphs of
all its builds.

-- Problem

We want an action scheduling algorithm that:

- can handle alternatives, aka "the GCC/libc situation"
- can determine scheduling feasibility ahead of time
- allows to schedule actions in parallel
- can handle inconsistent states
  - e.g. gcc is installed but libc is not

-- The GCC/libc situation

Builds can now depend on installation of any build from another
component.
We (rev.ng) use/plan to use this feature to:

- allow installing different builds of a dependency
- managing the dependency chain for bootstrapping cross compilers

By design, rev.ng orchestra configuration describes the dependency chain
for cross compilers with loops.
This is because when bootstrapping a compiler we want this dependency
chain:

```
gcc@stage2 -> libc@default -> gcc@stage1 -> libc@headers
```

Installing `stage2` causes `stage1` to be uninstalled, and installing
`libc@default` causes `libc@headers` to be uninstalled.  Because of this
if we used the above static dependency chain every time `gcc@stage2` is
needed we would need to reinstall everything from `libc@headers`
upwards.

Two ways to solve this problem are:

- "cutting" the dependency chain if `gcc@stage1` and/or `libc/headers`
  are not needed
- specifying that `libc@default` needs "any gcc" and that `gcc@stage1`
  needs "any libc", so that if `gcc@stage2` and `libc@default` are
  installed, we can satisfy these dependencies by having `gcc@stage2 ->
  libc@default` and `libc@default -> gcc@stage2`

The latter solution does imply generating a cyclic dependency cycle, but
this is done only if all its components are already installed
(recognizing there's nothing to do).

-- Solution

We first build an initial dependency graph, collecting all dependencies
for the "root" actions the user requested (e.g. install
`somecomponent`).  This graph contains some special "AnyOf" nodes which
represent alternatives.  Only one successor of these nodes will be
picked to satisfy the dependency.

The core step of the algorithm finds assignments for those nodes such
that either no cycles are found, or all nodes in the cycles are
satisfied.
The search is done such that already installed components and preferred
builds (`component~preferredbuild`) are tried first.
The algorithm assigns choices for different strongly connected
components of the dependency graph independently.
This optimization reduces the search space size from intractable to very
manageable (in my testing execution time goes from "never" finishing to
~0.1s for resolving `ui/cold-revng` dependency graph).

Then a couple of trivial simplification steps are performed on the
graph, and the actions are scheduled opportunistically in topological
order, allowing parallel execution.