Enforce ordering between same component actions
This commit solves an edge case with dependency graphs like

       +-------+
    +--+  A@1  +--+
    |  +-------+  |
+---v---+     +---v---+
|  B@1  |     |  A@2  +--+
+-------+     +-------+  |
                     +---v---+
                     |  B@2  |
                     +-------+

Where both following schedules would be possible:
- B@2, A@2, B@1, A@1
- B@2, B@1, A@2, A@1

Note that the second schedule runs B@1 after B@2 but before A@2,
which will not find the exact build it requested.
This should not be a problem if builds were exactly identical, but
this assumption is not true for rev.ng use case.
When bootstrapping a toolchain a dependency graph like this can be
generated (replace A with gcc stage1/2 and B with libc
default/headers)

A new pass is introduced which ensures that an action will always
find the exact builds it directly depends on installed.

This is done by transforming the graph into this (before transitive
reduction)

       +-------+
    +--+  A@1  +--+
    |  +-------+  |
+---v---+     +---v---+
|  B@1  +----->  A@2  +--+
+---+---+     +-------+  |
    |                +---v---+
    +---------------->  B@2  |
                     +-------+

See executor.py for a more in-depth explanation.