Create InstallMetadata class
The InstallMetadata class abstracts handling of the metadata that is
written to keep track of what is installed. This commit also moves
code handling the installed file indexes, which is not part of
InstallMetadata because we want to be able to handle them separately,
so that e.g. uninstalling a component is always possible even if the
other metadata is broken.