Allow debugging component hash computation
Warn: This commit bumps orchestra version because it changes the way a
component hash is computed.

Instead of computing a hash-of-hashes, now the hash is computed from a
single YAML document. This document is stored together with the binary
archive, and can be diffed against the current configuration hash
material to pinpoint which changes were made in case orchestra is
looking for a binary archive which does not exist.