executor: run actions depth-first
executor now runs available tasks in depth first.

This is achieved by sorting by component name the available tasks and
picking the first one.

This should avoid the situation, in particular in the CI, where we first
clone all components, then we configure them and finally install. This
should minimize the time to get to the first failure.